Cleanup, add a tool

+++ src-interface/viewer/image_handler.cpp
@@ -103,7 +103,7 @@ namespace satdump
                     [proj_func, corrected_stuff](float lat, float lon, int map_height, int map_width) mutable -> std::pair<int, int>
                 {
                     std::pair<int, int> ret = proj_func(lat, lon, map_height, map_width);
-                    if (ret.first != -1 && ret.second != -1 && ret.first < corrected_stuff.size() && ret.first >= 0)
+                    if (ret.first != -1 && ret.second != -1 && ret.first < (int)corrected_stuff.size() && ret.first >= 0)
                         ret.first = corrected_stuff[ret.first];
                     else
                         ret.second = ret.first = -1;
